requirements.txt
fastapi==0.115.11  # Основной фреймворк
uvicorn==0.34.0    # ASGI сервер
sqlalchemy[asyncio]==2.0.39  # ORM с поддержкой asyncio
asyncpg==0.29.0    # Драйвер для PostgreSQL
pydantic==2.10.6   # Валидация данных
email-validator==2.2.0  # Валидация email
python-jose[cryptography]==3.4.0  # JWT токены
bcrypt==4.0.1      # Хеширование паролей
python-dotenv==1.0.1  # Загрузка .env
gunicorn==23.0.0   # WSGI сервер (опционально, для продакшена)
alembic==1.13.3    # Миграции БД
pytest==8.3.3      # Тестирование
pytest-asyncio==0.24.0  # Поддержка asyncio в тестах
httpx==0.27.2      # HTTP-клиент для тестов
sqladmin==0.18.0  # Админ-панель для SQLAlchemy
jinja2==3.1.4      # Шаблонизатор для кастомных форм
python-multipart==0.0.19  # Поддержка multipart/form-data для загрузки файлов
wtforms==3.1.2     # Добавляем WTForms для обработки форм
itsdangerous==2.2.0
minio==7.2.15